refactor: rename searchParams to search-params and update references accordingly
